API: Move CYAML headers into cyaml directory.
