+++ test/units/test.c
@@ -12,7 +12,7 @@
 #include <string.h>
 #include <stdio.h>
 
-#include <cyaml.h>
+#include <cyaml/cyaml.h>
 
 #include "ttest.h"
